requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
oauthlib==3.2.2
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.4.0

pandas~=2.2.3

dotenv~=0.9.9
python-dotenv~=1.1.0

nltk~=3.9.1
textract~=1.6.5
openai~=1.78.1
